requirements.txt
aiogram==2.22.2
aiohttp==3.8.3
aiosignal==1.2.0
async-timeout==4.0.2
attrs==22.1.0
Babel==2.9.1
certifi==2022.9.24
charset-normalizer==2.1.1
frozenlist==1.3.1
idna==3.4
multidict==6.0.2
psycopg2-binary==2.9.4
python-dotenv==0.21.0
pytz==2022.5
requests==2.28.1
urllib3==1.26.12
yarl==1.8.1
